Refactor App and WebApp to centralize import and export functionality in AppMenu, removing redundant methods and enhancing code maintainability.
